Simplify leaderboard query
The old query did a join with "games_players where win" which meant that someone with a winratio of zero would not show up.

This could have been fixed by using a left join and a coalesce(..., 0).

However, it's actually possible to eliminate all the subqueries.

This new query also fixes the bug